Some methods in FileService were implemented
