+++ .idea/libraries/Maven__javax_ejb_javax_ejb_api_3_2.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: javax.ejb:javax.ejb-api:3.2">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/javax/ejb/javax.ejb-api/3.2/javax.ejb-api-3.2.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/javax/ejb/javax.ejb-api/3.2/javax.ejb-api-3.2-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/javax/ejb/javax.ejb-api/3.2/javax.ejb-api-3.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
