+++ JavaFiddle-Web/src/main/java/ru/javafiddle/web/services/ProjectService.java
@@ -1,11 +1,10 @@
-package services;
+package ru.javafiddle.web.services;
 
 import javax.ejb.EJB;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
